refactor: move helpers
