+++ src/lib/Scenes/SavedSearchAlert/helpers.ts
@@ -1,7 +1,10 @@
 import { defaultCommonFilterOptions, FilterParamName } from "lib/Components/ArtworkFilter/ArtworkFilterHelpers"
 import { SearchCriteriaAttributes } from "lib/Components/ArtworkFilter/SavedSearch/types"
+import { LegacyNativeModules } from "lib/NativeModules/LegacyNativeModules"
+import { getNotificationPermissionsStatus, PushAuthorizationStatus } from "lib/utils/PushNotification"
 import { groupBy } from "lodash"
 import { bullet } from "palette"
+import { Alert, AlertButton, Linking, Platform } from "react-native"
 import { extractPillValue } from "./pillExtractors"
 import { SavedSearchPill } from "./SavedSearchAlertModel"
 
@@ -37,4 +40,73 @@ export const getSearchCriteriaFromPills = (pills: SavedSearchPill[]) => {
   })
 
   return criteria
+}
+
+export const requestNotificationPermissions = () => {
+  // permissions not determined: Android should never need this
+  if (Platform.OS === "ios") {
+    Alert.alert(
+      "Artsy would like to send you notifications",
+      "We need your permission to send notifications on alerts you have created.",
+      [
+        {
+          text: "Proceed",
+          onPress: () => LegacyNativeModules.ARTemporaryAPIModule.requestDirectNotificationPermissions(),
+        },
+        {
+          text: "Cancel",
+          style: "cancel",
+        },
+      ]
+    )
+  }
+}
+
+export const showHowToEnableNotificationInstructionAlert = () => {
+  const deviceText = Platform.select({
+    ios: "iOS",
+    android: "android",
+    default: "device",
+  })
+  const instruction = Platform.select({
+    ios: `Tap 'Artsy' and enable "Allow Notifications" for Artsy.`,
+    default: "",
+  })
+
+  const buttons: AlertButton[] = [
+    {
+      text: "Settings",
+      onPress: () => {
+        if (Platform.OS === "android") {
+          Linking.openSettings()
+        } else {
+          Linking.openURL("App-prefs:NOTIFICATIONS_ID")
+        }
+      },
+    },
+    {
+      text: "Cancel",
+      style: "cancel",
+    },
+  ]
+
+  Alert.alert(
+    "Artsy would like to send you notifications",
+    `To receive notifications for your alerts, you will need to enable them in your ${deviceText} Settings. ${instruction}`,
+    Platform.OS === "ios" ? buttons : buttons.reverse()
+  )
+}
+
+export const checkOrRequestPushPermissions = async () => {
+  const notificationStatus = await getNotificationPermissionsStatus()
+
+  if (notificationStatus === PushAuthorizationStatus.Denied) {
+    showHowToEnableNotificationInstructionAlert()
+  }
+
+  if (notificationStatus === PushAuthorizationStatus.NotDetermined) {
+    requestNotificationPermissions()
+  }
+
+  return notificationStatus === PushAuthorizationStatus.Authorized
 }
